refactor(agent-core): modularize subprocess runtime
